--- xinference/web/ui/src/scenes/launch_model/styles/launchModelStyle.js
@@ -1,8 +1,0 @@
-const style = {
-  display: 'grid',
-  gridTemplateColumns: 'repeat(auto-fill, minmax(300px, 1fr))',
-  paddingLeft: '2rem',
-  gridGap: '2rem 0rem',
-}
-
-export default style
